+++ src/onegov.core/onegov/core/tests/test_layout.py
@@ -36,7 +36,6 @@ def test_format_date():
 
 
 def test_format_number():
-
     layout = Layout(model=object(), request=Bunch())
 
     layout.request.locale = 'de_CH'
@@ -45,6 +44,7 @@ def test_format_number():
     assert layout.format_number(1000) == "1’000"
     assert layout.format_number(1000.00) == "1’000.00"
     assert layout.format_number(1000.00, 0) == "1’000"
+    assert layout.format_number(None) == ""
 
     layout.request.locale = 'de'
     assert layout.format_number(100) == "100"
@@ -52,6 +52,7 @@ def test_format_number():
     assert layout.format_number(1000) == "1.000"
     assert layout.format_number(1000.00) == "1.000,00"
     assert layout.format_number(1000.00, 0) == "1.000"
+    assert layout.format_number(None) == ""
 
 
 def test_relative_date():